[update]修改了logo [update]修改了UserController.java中关于detail方法的身份判断逻辑 [update]修改了DTO中的ident字段为Byte和数据库统一 [update]修改了KaptchaController中的流write逻辑

+++ src/main/java/com/gdou/teaching/controller/common/UserController.java
@@ -116,9 +116,9 @@ public class UserController {
             UserDTO userByUserId = userService.getUserDetailByUserId(userId);
             UserVO userVO = new UserVO();
             BeanUtils.copyProperties(userByUserId, userVO);
-            if (userByUserId.getUserIdent().equals(UserIdentEnum.SUTUDENT.getCode())) {
+            if (userByUserId.getUserIdent().equals(UserIdentEnum.SUTUDENT.getCode().byteValue())) {
                 userVO.setUserIdent("学生");
-            } else if (userByUserId.getUserIdent().equals(UserIdentEnum.TEACHER.getCode())) {
+            } else if (userByUserId.getUserIdent().equals(UserIdentEnum.TEACHER.getCode().byteValue())) {
                 userVO.setUserIdent("老师");
             } else {
                 userVO.setUserIdent("管理员");
